replace the old codes
