+++ opentelekomcloud/provider.go
@@ -300,7 +300,6 @@ func Provider() terraform.ResourceProvider {
 			"opentelekomcloud_cce_node_v3":                        resourceCCENodeV3(),
 			"opentelekomcloud_cce_cluster_v3":                     resourceCCEClusterV3(),
 			"opentelekomcloud_maas_task_v1":                       resourceMaasTaskV1(),
-			"opentelekomcloud_identity_role":                      resourceIdentityRole(),
 		},
 
 		ConfigureFunc: configureProvider,
