+++ opentelekomcloud/resource_opentelekomcloud_identity_role_v3.go
@@ -1,11 +1,26 @@
+// ----------------------------------------------------------------------------
+//
+//     ***     AUTO GENERATED CODE    ***    AUTO GENERATED CODE     ***
+//
+// ----------------------------------------------------------------------------
+//
+//     This file is automatically generated by Magic Modules and manual
+//     changes will be clobbered when the file is regenerated.
+//
+//     Please read more about how to change this file at
+//     https://www.github.com/huaweicloud/magic-modules
+//
+// ----------------------------------------------------------------------------
+
 package opentelekomcloud
 
 import (
 	"fmt"
 	"log"
+	"reflect"
 
 	"github.com/hashicorp/terraform/helper/schema"
-	"github.com/huaweicloud/golangsdk/openstack/identity/v3/roles"
+	"github.com/huaweicloud/golangsdk"
 )
 
 func resourceIdentityRoleV3() *schema.Resource {
@@ -14,27 +29,75 @@ func resourceIdentityRoleV3() *schema.Resource {
 		Read:   resourceIdentityRoleV3Read,
 		Update: resourceIdentityRoleV3Update,
 		Delete: resourceIdentityRoleV3Delete,
+
 		Importer: &schema.ResourceImporter{
 			State: schema.ImportStatePassthrough,
 		},
 
 		Schema: map[string]*schema.Schema{
-			"domain_id": &schema.Schema{
+			"region": &schema.Schema{
 				Type:     schema.TypeString,
 				Optional: true,
+				Computed: true,
+			},
+
+			"description": &schema.Schema{
+				Type:     schema.TypeString,
+				Required: true,
+			},
+
+			"display_layer": &schema.Schema{
+				Type:     schema.TypeString,
+				Required: true,
+			},
+
+			"display_name": &schema.Schema{
+				Type:     schema.TypeString,
+				Required: true,
+			},
+
+			"policy": &schema.Schema{
+				Type:     schema.TypeList,
+				Required: true,
+				MaxItems: 1,
+				Elem: &schema.Resource{
+					Schema: map[string]*schema.Schema{
+						"statement": &schema.Schema{
+							Type:     schema.TypeList,
+							Required: true,
+							Elem: &schema.Resource{
+								Schema: map[string]*schema.Schema{
+									"action": &schema.Schema{
+										Type:     schema.TypeList,
+										Required: true,
+										Elem: &schema.Schema{
+											Type: schema.TypeString,
+										},
+									},
+									"effect": &schema.Schema{
+										Type:     schema.TypeString,
+										Required: true,
+									},
+								},
+							},
+						},
+					},
+				},
+			},
+
+			"catalog": &schema.Schema{
+				Type:     schema.TypeString,
+				Computed: true,
+			},
+
+			"domain_id": &schema.Schema{
+				Type:     schema.TypeString,
 				Computed: true,
 			},
 
 			"name": &schema.Schema{
 				Type:     schema.TypeString,
-				Required: true,
-			},
-
-			"region": &schema.Schema{
-				Type:     schema.TypeString,
-				Optional: true,
 				Computed: true,
-				ForceNew: true,
 			},
 		},
 	}
@@ -42,68 +105,152 @@ func resourceIdentityRoleV3() *schema.Resource {
 
 func resourceIdentityRoleV3Create(d *schema.ResourceData, meta interface{}) error {
 	config := meta.(*Config)
-	identityClient, err := config.identityV3Client(GetRegion(d, config))
+	client, err := config.sdkClient("", "identity", serviceDomainLevel)
 	if err != nil {
-		return fmt.Errorf("Error creating OpenStack identity client: %s", err)
+		return fmt.Errorf("Error creating sdk client, err=%s", err)
 	}
+	client.Endpoint = "https://iam.eu-de.otc.t-systems.com/v3.0/"
 
-	createOpts := roles.CreateOpts{
-		DomainID: d.Get("domain_id").(string),
-		Name:     d.Get("name").(string),
-	}
+	opts := make(map[string]interface{})
+	opts["description"] = d.Get("description")
+	opts["display_layer"] = d.Get("display_layer")
+	opts["display_name"] = d.Get("display_name")
+	opts["policy"] = d.Get("policy")
 
-	log.Printf("[DEBUG] Create Options: %#v", createOpts)
-	role, err := roles.Create(identityClient, createOpts).Extract()
+	params := make(map[string]interface{})
+
+	roleProp, err := expandIdentityRoleV3CreateRole(opts, nil)
 	if err != nil {
-		return fmt.Errorf("Error creating OpenStack role: %s", err)
+		return err
+	}
+	e, err := isEmptyValue(reflect.ValueOf(roleProp))
+	if err != nil {
+		return err
+	}
+	if !e {
+		params["role"] = roleProp
 	}
 
-	d.SetId(role.ID)
+	log.Printf("[DEBUG] Creating new Role: %#v", params)
+
+	url, err := replaceVars(d, "OS-ROLE/roles", nil)
+	if err != nil {
+		return err
+	}
+	url = client.ServiceURL(url)
+
+	r := golangsdk.Result{}
+	_, r.Err = client.Post(url, &params, &r.Body, &golangsdk.RequestOpts{
+		OkCodes: successHTTPCodes,
+	})
+	if r.Err != nil {
+		return fmt.Errorf("Error creating Role: %s", r.Err)
+	}
+
+	id, err := navigateValue(r.Body, []string{"role", "id"}, nil)
+	if err != nil {
+		return fmt.Errorf("Error constructing id: %s", err)
+	}
+	d.SetId(id.(string))
 
 	return resourceIdentityRoleV3Read(d, meta)
 }
 
 func resourceIdentityRoleV3Read(d *schema.ResourceData, meta interface{}) error {
 	config := meta.(*Config)
-	identityClient, err := config.identityV3Client(GetRegion(d, config))
+	client, err := config.sdkClient("", "identity", serviceDomainLevel)
 	if err != nil {
-		return fmt.Errorf("Error creating OpenStack identity client: %s", err)
+		return fmt.Errorf("Error creating sdk client, err=%s", err)
+	}
+	client.Endpoint = "https://iam.eu-de.otc.t-systems.com/v3.0/"
+
+	url, err := replaceVars(d, "OS-ROLE/roles/{id}", nil)
+	if err != nil {
+		return err
+	}
+	url = client.ServiceURL(url)
+
+	r := golangsdk.Result{}
+	_, r.Err = client.Get(
+		url, &r.Body,
+		&golangsdk.RequestOpts{MoreHeaders: map[string]string{"Content-Type": "application/json"}})
+	if r.Err != nil {
+		return fmt.Errorf("Error reading %s: %s", fmt.Sprintf("IdentityRoleV3 %q", d.Id()), r.Err)
+	}
+	res, ok := r.Body.(map[string]interface{})
+	if !ok {
+		return fmt.Errorf("Error reading %s: the result is not map", fmt.Sprintf("IdentityRoleV3 %q", d.Id()))
 	}
 
-	role, err := roles.Get(identityClient, d.Id()).Extract()
+	catalogProp, err := navigateValue(res, []string{"role", "catalog"}, nil)
 	if err != nil {
-		return CheckDeleted(d, err, "role")
+		return fmt.Errorf("Error reading Role:catalog, err: %s", err)
+	}
+	if err := d.Set("catalog", catalogProp); err != nil {
+		return fmt.Errorf("Error setting Role:catalog, err: %s", err)
 	}
 
-	log.Printf("[DEBUG] Retrieved OpenStack role: %#v", role)
+	domainIDProp, err := navigateValue(res, []string{"role", "domain_id"}, nil)
+	if err != nil {
+		return fmt.Errorf("Error reading Role:domain_id, err: %s", err)
+	}
+	if err := d.Set("domain_id", domainIDProp); err != nil {
+		return fmt.Errorf("Error setting Role:domain_id, err: %s", err)
+	}
 
-	d.Set("domain_id", role.DomainID)
-	d.Set("name", role.Name)
-	d.Set("region", GetRegion(d, config))
+	nameProp, err := navigateValue(res, []string{"role", "name"}, nil)
+	if err != nil {
+		return fmt.Errorf("Error reading Role:name, err: %s", err)
+	}
+	if err := d.Set("name", nameProp); err != nil {
+		return fmt.Errorf("Error setting Role:name, err: %s", err)
+	}
 
 	return nil
 }
 
 func resourceIdentityRoleV3Update(d *schema.ResourceData, meta interface{}) error {
 	config := meta.(*Config)
-	identityClient, err := config.identityV3Client(GetRegion(d, config))
+	client, err := config.sdkClient("", "identity", serviceDomainLevel)
 	if err != nil {
-		return fmt.Errorf("Error creating OpenStack identity client: %s", err)
+		return fmt.Errorf("Error creating sdk client, err=%s", err)
+	}
+	client.Endpoint = "https://iam.eu-de.otc.t-systems.com/v3.0/"
+
+	opts := make(map[string]interface{})
+	opts["description"] = d.Get("description")
+	opts["display_layer"] = d.Get("display_layer")
+	opts["display_name"] = d.Get("display_name")
+	opts["policy"] = d.Get("policy")
+
+	params := make(map[string]interface{})
+
+	roleProp, err := expandIdentityRoleV3UpdateRole(opts, nil)
+	if err != nil {
+		return err
+	}
+	e, err := isEmptyValue(reflect.ValueOf(roleProp))
+	if err != nil {
+		return err
+	}
+	if !e {
+		params["role"] = roleProp
 	}
 
-	var hasChange bool
-	var updateOpts roles.UpdateOpts
+	log.Printf("[DEBUG] Updating Role %q: %#v", d.Id(), params)
 
-	if d.HasChange("name") {
-		hasChange = true
-		updateOpts.Name = d.Get("name").(string)
+	url, err := replaceVars(d, "OS-ROLE/roles/{id}", nil)
+	if err != nil {
+		return err
 	}
+	url = client.ServiceURL(url)
 
-	if hasChange {
-		_, err := roles.Update(identityClient, d.Id(), updateOpts).Extract()
-		if err != nil {
-			return fmt.Errorf("Error updating OpenStack role: %s", err)
-		}
+	r := golangsdk.Result{}
+	_, r.Err = client.Patch(url, &params, &r.Body, &golangsdk.RequestOpts{
+		OkCodes: successHTTPCodes,
+	})
+	if r.Err != nil {
+		return fmt.Errorf("Error updating Role %q: %s", d.Id(), r.Err)
 	}
 
 	return resourceIdentityRoleV3Read(d, meta)
@@ -111,15 +258,318 @@ func resourceIdentityRoleV3Update(d *schema.ResourceData, meta interface{}) erro
 
 func resourceIdentityRoleV3Delete(d *schema.ResourceData, meta interface{}) error {
 	config := meta.(*Config)
-	identityClient, err := config.identityV3Client(GetRegion(d, config))
+	client, err := config.sdkClient("", "identity", serviceDomainLevel)
 	if err != nil {
-		return fmt.Errorf("Error creating OpenStack identity client: %s", err)
+		return fmt.Errorf("Error creating sdk client, err=%s", err)
 	}
+	client.Endpoint = "https://iam.eu-de.otc.t-systems.com/v3.0/"
 
-	err = roles.Delete(identityClient, d.Id()).ExtractErr()
+	url, err := replaceVars(d, "OS-ROLE/roles/{id}", nil)
 	if err != nil {
-		return fmt.Errorf("Error deleting OpenStack role: %s", err)
+		return err
+	}
+	url = client.ServiceURL(url)
+
+	log.Printf("[DEBUG] Deleting Role %q", d.Id())
+	r := golangsdk.Result{}
+	_, r.Err = client.Delete(url, &golangsdk.RequestOpts{
+		OkCodes:      successHTTPCodes,
+		JSONResponse: &r.Body,
+		MoreHeaders:  map[string]string{"Content-Type": "application/json"},
+	})
+	if r.Err != nil {
+		return fmt.Errorf("Error deleting Role %q: %s", d.Id(), r.Err)
 	}
 
 	return nil
+}
+
+func expandIdentityRoleV3CreateRole(d interface{}, arrayIndex map[string]int) (interface{}, error) {
+	req := make(map[string]interface{})
+
+	descriptionProp, err := navigateValue(d, []string{"description"}, arrayIndex)
+	if err != nil {
+		return nil, err
+	}
+	e, err := isEmptyValue(reflect.ValueOf(descriptionProp))
+	if err != nil {
+		return nil, err
+	}
+	if !e {
+		req["description"] = descriptionProp
+	}
+
+	displayNameProp, err := navigateValue(d, []string{"display_name"}, arrayIndex)
+	if err != nil {
+		return nil, err
+	}
+	e, err = isEmptyValue(reflect.ValueOf(displayNameProp))
+	if err != nil {
+		return nil, err
+	}
+	if !e {
+		req["display_name"] = displayNameProp
+	}
+
+	var localArrayIndex = make(map[string]int)
+
+	if arrayIndex != nil {
+		for k, v := range arrayIndex {
+			localArrayIndex[k] = v
+		}
+	}
+	localArrayIndex["policy"] = 0
+
+	policyProp, err := expandIdentityRoleV3CreateRolePolicy(d, localArrayIndex)
+	if err != nil {
+		return nil, err
+	}
+	e, err = isEmptyValue(reflect.ValueOf(policyProp))
+	if err != nil {
+		return nil, err
+	}
+	if !e {
+		req["policy"] = policyProp
+	}
+
+	typeProp, err := expandIdentityRoleV3CreateRoleType(d, arrayIndex)
+	if err != nil {
+		return nil, err
+	}
+	e, err = isEmptyValue(reflect.ValueOf(typeProp))
+	if err != nil {
+		return nil, err
+	}
+	if !e {
+		req["type"] = typeProp
+	}
+
+	return req, nil
+}
+
+func expandIdentityRoleV3CreateRolePolicy(d interface{}, arrayIndex map[string]int) (interface{}, error) {
+	req := make(map[string]interface{})
+
+	statementProp, err := expandIdentityRoleV3CreateRolePolicyStatement(d, arrayIndex)
+	if err != nil {
+		return nil, err
+	}
+	e, err := isEmptyValue(reflect.ValueOf(statementProp))
+	if err != nil {
+		return nil, err
+	}
+	if !e {
+		req["Statement"] = statementProp
+	}
+
+	req["Version"] = "1.1"
+
+	return req, nil
+}
+
+func expandIdentityRoleV3CreateRolePolicyStatement(d interface{}, arrayIndex map[string]int) (interface{}, error) {
+	newArrayIndex := make(map[string]int)
+	if arrayIndex != nil {
+		for k, v := range arrayIndex {
+			newArrayIndex[k] = v
+		}
+	}
+
+	v, err := navigateValue(d, []string{"policy", "statement"}, newArrayIndex)
+	if err != nil {
+		return nil, err
+	}
+
+	n := len(v.([]interface{}))
+	req := make([]interface{}, 0, n)
+	for i := 0; i < n; i++ {
+		newArrayIndex["statement"] = i
+		transformed := make(map[string]interface{})
+
+		actionProp, err := navigateValue(d, []string{"policy", "statement", "action"}, newArrayIndex)
+		if err != nil {
+			return nil, err
+		}
+		e, err := isEmptyValue(reflect.ValueOf(actionProp))
+		if err != nil {
+			return nil, err
+		}
+		if !e {
+			transformed["Action"] = actionProp
+		}
+
+		effectProp, err := navigateValue(d, []string{"policy", "statement", "effect"}, newArrayIndex)
+		if err != nil {
+			return nil, err
+		}
+		e, err = isEmptyValue(reflect.ValueOf(effectProp))
+		if err != nil {
+			return nil, err
+		}
+		if !e {
+			transformed["Effect"] = effectProp
+		}
+
+		req = append(req, transformed)
+	}
+
+	return req, nil
+}
+
+func expandIdentityRoleV3CreateRoleType(d interface{}, arrayIndex map[string]int) (interface{}, error) {
+	v, err := navigateValue(d, []string{"display_layer"}, arrayIndex)
+	if err != nil {
+		return nil, err
+	}
+	if v == "domain" {
+		return "AX", nil
+	} else if v == "project" {
+		return "XA", nil
+	}
+	return nil, fmt.Errorf("unknown display layer:%v", v)
+}
+
+func expandIdentityRoleV3UpdateRole(d interface{}, arrayIndex map[string]int) (interface{}, error) {
+	req := make(map[string]interface{})
+
+	descriptionProp, err := navigateValue(d, []string{"description"}, arrayIndex)
+	if err != nil {
+		return nil, err
+	}
+	e, err := isEmptyValue(reflect.ValueOf(descriptionProp))
+	if err != nil {
+		return nil, err
+	}
+	if !e {
+		req["description"] = descriptionProp
+	}
+
+	displayNameProp, err := navigateValue(d, []string{"display_name"}, arrayIndex)
+	if err != nil {
+		return nil, err
+	}
+	e, err = isEmptyValue(reflect.ValueOf(displayNameProp))
+	if err != nil {
+		return nil, err
+	}
+	if !e {
+		req["display_name"] = displayNameProp
+	}
+
+	var localArrayIndex = make(map[string]int)
+
+	if arrayIndex != nil {
+		for k, v := range arrayIndex {
+			localArrayIndex[k] = v
+		}
+	}
+	localArrayIndex["policy"] = 0
+
+	policyProp, err := expandIdentityRoleV3UpdateRolePolicy(d, localArrayIndex)
+	if err != nil {
+		return nil, err
+	}
+	e, err = isEmptyValue(reflect.ValueOf(policyProp))
+	if err != nil {
+		return nil, err
+	}
+	if !e {
+		req["policy"] = policyProp
+	}
+
+	typeProp, err := expandIdentityRoleV3UpdateRoleType(d, arrayIndex)
+	if err != nil {
+		return nil, err
+	}
+	e, err = isEmptyValue(reflect.ValueOf(typeProp))
+	if err != nil {
+		return nil, err
+	}
+	if !e {
+		req["type"] = typeProp
+	}
+
+	return req, nil
+}
+
+func expandIdentityRoleV3UpdateRolePolicy(d interface{}, arrayIndex map[string]int) (interface{}, error) {
+	req := make(map[string]interface{})
+
+	statementProp, err := expandIdentityRoleV3UpdateRolePolicyStatement(d, arrayIndex)
+	if err != nil {
+		return nil, err
+	}
+	e, err := isEmptyValue(reflect.ValueOf(statementProp))
+	if err != nil {
+		return nil, err
+	}
+	if !e {
+		req["Statement"] = statementProp
+	}
+
+	req["Version"] = "1.1"
+
+	return req, nil
+}
+
+func expandIdentityRoleV3UpdateRolePolicyStatement(d interface{}, arrayIndex map[string]int) (interface{}, error) {
+	newArrayIndex := make(map[string]int)
+	if arrayIndex != nil {
+		for k, v := range arrayIndex {
+			newArrayIndex[k] = v
+		}
+	}
+
+	v, err := navigateValue(d, []string{"policy", "statement"}, newArrayIndex)
+	if err != nil {
+		return nil, err
+	}
+
+	n := len(v.([]interface{}))
+	req := make([]interface{}, 0, n)
+	for i := 0; i < n; i++ {
+		newArrayIndex["statement"] = i
+		transformed := make(map[string]interface{})
+
+		actionProp, err := navigateValue(d, []string{"policy", "statement", "action"}, newArrayIndex)
+		if err != nil {
+			return nil, err
+		}
+		e, err := isEmptyValue(reflect.ValueOf(actionProp))
+		if err != nil {
+			return nil, err
+		}
+		if !e {
+			transformed["Action"] = actionProp
+		}
+
+		effectProp, err := navigateValue(d, []string{"policy", "statement", "effect"}, newArrayIndex)
+		if err != nil {
+			return nil, err
+		}
+		e, err = isEmptyValue(reflect.ValueOf(effectProp))
+		if err != nil {
+			return nil, err
+		}
+		if !e {
+			transformed["Effect"] = effectProp
+		}
+
+		req = append(req, transformed)
+	}
+
+	return req, nil
+}
+
+func expandIdentityRoleV3UpdateRoleType(d interface{}, arrayIndex map[string]int) (interface{}, error) {
+	v, err := navigateValue(d, []string{"display_layer"}, arrayIndex)
+	if err != nil {
+		return nil, err
+	}
+	if v == "domain" {
+		return "AX", nil
+	} else if v == "project" {
+		return "XA", nil
+	}
+	return nil, fmt.Errorf("unknown display layer:%v", v)
 }
