+++ website/opentelekomcloud.erb
@@ -232,8 +232,8 @@
             <li<%= sidebar_current("docs-opentelekomcloud-resource-identity-role-assignment-v3") %>>
               <a href="/docs/providers/opentelekomcloud/r/identity_role_assignment_v3.html">opentelekomcloud_identity_role_assignment_v3</a>
             </li>
-            <li<%= sidebar_current("docs-opentelekomcloud-resource-identity-role") %>>
-              <a href="/docs/providers/opentelekomcloud/r/identity_role.html">opentelekomcloud_identity_role</a>
+            <li<%= sidebar_current("docs-opentelekomcloud-resource-identity-role-v3") %>>
+              <a href="/docs/providers/opentelekomcloud/r/identity_role_v3.html">opentelekomcloud_identity_role_v3</a>
             </li>
           </ul>
         </li>
